+++ pong-game/components/Friends/FriendsMainContents.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react'
-import styles from '@/styles/components/Friends/FriendsMainContents.module.css'
+import styles from './FriendsMainContents.module.css'
 import FriendsNaviTab from './FriendsNaviTab'
 import Image from 'next/image'
 import profileImage from '@/public/img/chat/userProfileImage.svg'
