+++ pong-game/components/dropDown/DropDown.tsx
@@ -1,5 +1,5 @@
-import styles1 from '@/styles/components/dropDown/DropDown.module.scss'
-import styles from '@/styles/rank/Rank.module.scss'
+import styles1 from './DropDown.module.scss'
+import styles from '../../pages/rank/rank.module.scss'
 import DropDownUserProfile from './DropDownUserProfile'
 
 interface dropDownProps {
