+++ pong-game/function/game/customRadio.tsx
@@ -4,8 +4,7 @@ import normalBtn from '../../public/img/game/normal.svg'
 import selectSpecialBtn from '../../public/img/game/selectSpecial.svg'
 import selectNormalBtn from '../../public/img/game/selectNormal.svg'
 import Image from 'next/image'
-// import styles from '../../styles/components/game/normalGame.module.scss'
-import styles from '../../styles/function/game/customRadio.module.scss'
+import styles from './customRadio.module.scss'
 
 interface radioOption{
   value: string;
