+++ pong-game/pages/main/index.tsx
@@ -1,6 +1,4 @@
-import Image from 'next/image'
-
-import styles from '../../styles/main/index.module.scss'
+import styles from './index.module.scss'
 import Layout from '@/components/layout/Layout'
 import { useState } from 'react'
 import NomalGame from '@/components/game/NormalGame'
