+++ pong-game/pages/rank/index.tsx
@@ -1,5 +1,5 @@
 import PageTitle from '@/components/UI/PageTitle'
-import styles from '../../styles/rank/Rank.module.scss'
+import styles from './Rank.module.scss'
 import profileImage from '../../public/img/mypage/profileImage.svg'
 import Image from 'next/image'
 import DropDown from '@/components/dropDown/DropDown'
